pyproject.toml
[tool.poetry]
name = "cwcom"
version = "2.0.0"
description = "carolinewoolard.com"
authors = ["Julian Boilen <julianboilen@gmail.com>", "Or Zubalsky <orzubalsky@gmail.com>"]
readme = "README.md"
packages = [
  {include = "portfolio"}
]

[tool.poetry.dependencies]
python = "~3.11"
django = "^5.1.4"
Pillow = "^11.1.0"
django-tinymce = "^4.1.0"
psycopg2-binary = "^2.9.5"
django-storages = "^1.13.1"
django-clearcache = "^1.2.1"
boto3 = "^1.35.90"
Werkzeug = "^2.2.2"
django-s3-upload = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
